refactor: builder transactions
