--- crates/op-rbuilder/src/builder/builder_tx.rs
@@ -1,695 +1,0 @@
-use alloy_consensus::TxEip1559;
-use alloy_eips::{Encodable2718, eip7623::TOTAL_COST_FLOOR_PER_TOKEN};
-use alloy_evm::{Database, rpc::TryIntoTxEnv};
-use alloy_op_evm::{OpEvm, OpTx};
-use alloy_primitives::{Address, B256, BlockHash, Bytes, TxKind, U256, map::HashSet};
-use alloy_sol_types::{ContractError, Revert, SolCall, SolError, SolInterface};
-use core::fmt::Debug;
-use op_alloy_consensus::OpTypedTransaction;
-use op_alloy_rpc_types::OpTransactionRequest;
-use op_revm::{OpHaltReason, OpTransactionError};
-use reth_evm::{Evm, EvmError, InvalidTxError, precompiles::PrecompilesMap};
-use reth_node_api::PayloadBuilderError;
-use reth_optimism_primitives::OpTransactionSigned;
-use reth_primitives_traits::Recovered;
-use reth_provider::{ProviderError, StateProvider};
-use reth_revm::{State, database::StateProviderDatabase};
-use reth_rpc_api::eth::EthTxEnvError;
-use revm::{
-    DatabaseCommit, DatabaseRef,
-    context::result::{EVMError, ExecutionResult, ResultAndState},
-    inspector::NoOpInspector,
-};
-use tracing::{error, trace, warn};
-
-use crate::{
-    evm::OpBlockEvmFactory, hardforks::ActiveHardforks, primitives::reth::ExecutionInfo,
-    tx_signer::Signer,
-};
-
-#[derive(Debug, Default)]
-pub struct SimulationSuccessResult<T: SolCall> {
-    pub gas_used: u64,
-    pub output: T::Return,
-    pub state_changes: revm::state::EvmState,
-}
-
-#[derive(Debug, Clone)]
-pub struct BuilderTransactionCtx {
-    pub gas_used: u64,
-    pub da_size: u64,
-    pub signed_tx: Recovered<OpTransactionSigned>,
-    // whether the transaction should be a top of block or
-    // bottom of block transaction
-    pub is_top_of_block: bool,
-}
-
-impl BuilderTransactionCtx {
-    pub fn set_top_of_block(mut self) -> Self {
-        self.is_top_of_block = true;
-        self
-    }
-
-    pub fn set_bottom_of_block(mut self) -> Self {
-        self.is_top_of_block = false;
-        self
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum InvalidContractDataError {
-    #[error("did not find expected logs expected {0:?} but got {1:?}")]
-    InvalidLogs(Vec<B256>, Vec<B256>),
-    #[error("could not decode output from contract call")]
-    OutputAbiDecodeError,
-}
-
-/// Possible error variants during construction of builder txs.
-#[derive(Debug, thiserror::Error)]
-pub enum BuilderTransactionError {
-    /// Builder account load fails to get builder nonce
-    #[error("failed to load account {0}")]
-    AccountLoadFailed(Address),
-    /// Signature signing fails
-    #[error("failed to sign transaction: {0}")]
-    SigningError(secp256k1::Error),
-    /// Invalid contract errors indicating the contract is incorrect
-    #[error("contract {0} may be incorrect, invalid contract data: {1}")]
-    InvalidContract(Address, InvalidContractDataError),
-    /// Transaction halted execution
-    #[error("transaction to {0} halted {1:?}")]
-    TransactionHalted(Address, OpHaltReason),
-    /// Transaction reverted
-    #[error("transaction to {0} reverted {1}")]
-    TransactionReverted(Address, Revert),
-    /// Invalid tx errors during evm execution.
-    #[error("invalid transaction error {0}")]
-    InvalidTransactionError(Box<dyn core::error::Error + Send + Sync>),
-    /// Unrecoverable error during evm execution.
-    #[error("evm execution error {0}")]
-    EvmExecutionError(Box<dyn core::error::Error + Send + Sync>),
-    /// Any other builder transaction errors.
-    #[error(transparent)]
-    Other(Box<dyn core::error::Error + Send + Sync>),
-}
-
-impl From<secp256k1::Error> for BuilderTransactionError {
-    fn from(error: secp256k1::Error) -> Self {
-        BuilderTransactionError::SigningError(error)
-    }
-}
-
-impl From<EVMError<ProviderError, OpTransactionError>> for BuilderTransactionError {
-    fn from(error: EVMError<ProviderError, OpTransactionError>) -> Self {
-        BuilderTransactionError::EvmExecutionError(Box::new(error))
-    }
-}
-
-impl From<EthTxEnvError> for BuilderTransactionError {
-    fn from(error: EthTxEnvError) -> Self {
-        BuilderTransactionError::EvmExecutionError(Box::new(error))
-    }
-}
-
-impl From<BuilderTransactionError> for PayloadBuilderError {
-    fn from(error: BuilderTransactionError) -> Self {
-        match error {
-            BuilderTransactionError::EvmExecutionError(e) => {
-                PayloadBuilderError::EvmExecutionError(e)
-            }
-            _ => PayloadBuilderError::other(error),
-        }
-    }
-}
-
-impl BuilderTransactionError {
-    pub fn other(error: impl core::error::Error + Send + Sync + 'static) -> Self {
-        BuilderTransactionError::Other(Box::new(error))
-    }
-
-    pub fn msg(msg: impl core::fmt::Display) -> Self {
-        Self::Other(msg.to_string().into())
-    }
-}
-
-/// Minimal environment needed by [`BuilderTransactions`] so that
-/// builder-transaction code does not depend on the full payload-job context.
-pub struct BuilderTxEnv<'a> {
-    pub evm_factory: &'a OpBlockEvmFactory,
-    pub hardforks: &'a ActiveHardforks,
-    pub base_fee: u64,
-    pub block_number: u64,
-    pub block_gas_limit: u64,
-    pub parent_hash: BlockHash,
-    pub max_uncompressed_block_size: Option<u64>,
-}
-
-impl BuilderTxEnv<'_> {
-    pub fn chain_id(&self) -> u64 {
-        self.hardforks.chain_id()
-    }
-}
-
-pub trait BuilderTransactions {
-    // Simulates and returns the signed builder transactions. The simulation modifies and commit
-    // changes to the db so call new_simulation_state to simulate on a new copy of the state
-    #[expect(clippy::too_many_arguments)]
-    fn simulate_builder_txs(
-        &self,
-        state_provider: impl StateProvider + Clone,
-        info: &mut ExecutionInfo,
-        ctx: &BuilderTxEnv<'_>,
-        db: &mut State<impl Database + DatabaseRef>,
-        top_of_block: bool,
-        is_first_flashblock: bool,
-        is_last_flashblock: bool,
-    ) -> Result<Vec<BuilderTransactionCtx>, BuilderTransactionError>;
-
-    #[expect(clippy::too_many_arguments)]
-    fn simulate_builder_txs_with_state_copy(
-        &self,
-        state_provider: impl StateProvider + Clone,
-        info: &mut ExecutionInfo,
-        ctx: &BuilderTxEnv<'_>,
-        db: &State<impl Database>,
-        top_of_block: bool,
-        is_first_flashblock: bool,
-        is_last_flashblock: bool,
-    ) -> Result<Vec<BuilderTransactionCtx>, BuilderTransactionError> {
-        let mut simulation_state = self.new_simulation_state(state_provider.clone(), db);
-        self.simulate_builder_txs(
-            state_provider,
-            info,
-            ctx,
-            &mut simulation_state,
-            top_of_block,
-            is_first_flashblock,
-            is_last_flashblock,
-        )
-    }
-
-    #[expect(clippy::too_many_arguments)]
-    fn add_builder_txs(
-        &self,
-        state_provider: impl StateProvider + Clone,
-        info: &mut ExecutionInfo,
-        ctx: &BuilderTxEnv<'_>,
-        db: &mut State<impl Database>,
-        top_of_block: bool,
-        is_first_flashblock: bool,
-        is_last_flashblock: bool,
-    ) -> Result<Vec<BuilderTransactionCtx>, BuilderTransactionError> {
-        let builder_txs = self.simulate_builder_txs_with_state_copy(
-            state_provider,
-            info,
-            ctx,
-            db,
-            top_of_block,
-            is_first_flashblock,
-            is_last_flashblock,
-        )?;
-
-        let mut evm = ctx.evm_factory.evm(&mut *db);
-
-        let mut invalid = HashSet::new();
-
-        for builder_tx in builder_txs.iter() {
-            if builder_tx.is_top_of_block != top_of_block {
-                // don't commit tx if the buidler tx is not being added in the intended
-                // position in the block
-                continue;
-            }
-            if invalid.contains(&builder_tx.signed_tx.signer()) {
-                warn!(
-                    target: "payload_builder",
-                    tx_hash = %builder_tx.signed_tx.tx_hash(),
-                    "builder signer invalid as previous builder tx reverted"
-                );
-                continue;
-            }
-
-            let ResultAndState { result, state } = match evm.transact(&builder_tx.signed_tx) {
-                Ok(res) => res,
-                Err(err) => {
-                    if let Some(err) = err.as_invalid_tx_err() {
-                        if err.is_nonce_too_low() {
-                            // if the nonce is too low, we can skip this transaction
-                            trace!(
-                                target: "payload_builder",
-                                error = %err,
-                                tx_hash = %builder_tx.signed_tx.tx_hash(),
-                                "skipping nonce too low builder transaction"
-                            );
-                        } else {
-                            // if the transaction is invalid, we can skip it and all of its
-                            // descendants
-                            trace!(
-                                target: "payload_builder",
-                                error = %err,
-                                tx_hash = %builder_tx.signed_tx.tx_hash(),
-                                "skipping invalid builder transaction and its descendants"
-                            );
-                            invalid.insert(builder_tx.signed_tx.signer());
-                        }
-
-                        continue;
-                    }
-                    // this is an error that we should treat as fatal for this attempt
-                    return Err(BuilderTransactionError::EvmExecutionError(Box::new(err)));
-                }
-            };
-
-            if !result.is_success() {
-                warn!(
-                    target: "payload_builder",
-                    tx_hash = %builder_tx.signed_tx.tx_hash(),
-                    result = ?result,
-                    "builder tx reverted"
-                );
-                invalid.insert(builder_tx.signed_tx.signer());
-                continue;
-            }
-
-            let tx_uncompressed_size = builder_tx.signed_tx.inner().encode_2718_len() as u64;
-
-            // Skip the builder tx if including it would push the cumulative
-            // uncompressed block size over the configured limit. The state
-            // changes from the simulation are dropped (we never call commit).
-            if let Some(limit) = ctx.max_uncompressed_block_size
-                && info.cumulative_uncompressed_bytes + tx_uncompressed_size > limit
-            {
-                warn!(
-                    target: "payload_builder",
-                    tx_hash = %builder_tx.signed_tx.tx_hash(),
-                    cumulative_uncompressed = info.cumulative_uncompressed_bytes,
-                    tx_uncompressed_size,
-                    limit,
-                    "skipping builder tx: would exceed max uncompressed block size"
-                );
-                continue;
-            }
-
-            info.commit_tx(
-                &builder_tx.signed_tx,
-                result,
-                state,
-                builder_tx.da_size,
-                None,
-                None,
-                ctx.evm_factory,
-                ctx.hardforks,
-                &mut evm,
-            );
-        }
-
-        Ok(builder_txs)
-    }
-
-    // Creates a copy of the state to simulate against
-    fn new_simulation_state(
-        &self,
-        state_provider: impl StateProvider,
-        db: &State<impl Database>,
-    ) -> State<StateProviderDatabase<impl StateProvider>> {
-        let state = StateProviderDatabase::new(state_provider);
-
-        State::builder()
-            .with_database(state)
-            .with_cached_prestate(db.cache.clone())
-            .with_bundle_update()
-            .build()
-    }
-
-    fn sign_tx(
-        &self,
-        to: Address,
-        from: Signer,
-        gas_used: u64,
-        calldata: Bytes,
-        ctx: &BuilderTxEnv<'_>,
-        db: impl DatabaseRef,
-    ) -> Result<Recovered<OpTransactionSigned>, BuilderTransactionError> {
-        let nonce = get_nonce(db, from.address)?;
-        // Create the EIP-1559 transaction
-        let tx = OpTypedTransaction::Eip1559(TxEip1559 {
-            chain_id: ctx.chain_id(),
-            nonce,
-            // Due to EIP-150, 63/64 of available gas is forwarded to external calls so need to add a buffer
-            gas_limit: gas_used * 64 / 63,
-            max_fee_per_gas: ctx.base_fee.into(),
-            to: TxKind::Call(to),
-            input: calldata,
-            ..Default::default()
-        });
-        Ok(from.sign_tx(tx)?)
-    }
-
-    fn commit_txs(
-        &self,
-        signed_txs: Vec<Recovered<OpTransactionSigned>>,
-        ctx: &BuilderTxEnv<'_>,
-        db: &mut State<impl Database>,
-    ) -> Result<(), BuilderTransactionError> {
-        let mut evm = ctx.evm_factory.evm(&mut *db);
-        for signed_tx in signed_txs {
-            let ResultAndState { state, .. } = evm
-                .transact(&signed_tx)
-                .map_err(|err| BuilderTransactionError::EvmExecutionError(Box::new(err)))?;
-            evm.db_mut().commit(state)
-        }
-        Ok(())
-    }
-
-    fn simulate_call<T: SolCall, E: SolInterface + Debug>(
-        &self,
-        tx: OpTransactionRequest,
-        expected_logs: Vec<B256>,
-        evm: &mut OpEvm<impl Database, NoOpInspector, PrecompilesMap>,
-    ) -> Result<SimulationSuccessResult<T>, BuilderTransactionError> {
-        let evm_env = alloy_evm::EvmEnv::new(evm.cfg_env().clone(), evm.block().clone());
-        let tx_env: revm::context::TxEnv = tx.as_ref().clone().try_into_tx_env(&evm_env)?;
-        let to = tx_env.kind.into_to().unwrap_or_default();
-        let op_tx = OpTx(op_revm::OpTransaction {
-            base: tx_env,
-            enveloped_tx: Some(Bytes::new()),
-            deposit: Default::default(),
-        });
-
-        let ResultAndState { result, state } = match evm.transact(op_tx) {
-            Ok(res) => res,
-            Err(err) => {
-                if err.is_invalid_tx_err() {
-                    return Err(BuilderTransactionError::InvalidTransactionError(Box::new(
-                        err,
-                    )));
-                } else {
-                    return Err(BuilderTransactionError::EvmExecutionError(Box::new(err)));
-                }
-            }
-        };
-        let gas_used = result.tx_gas_used();
-
-        match result {
-            ExecutionResult::Success { output, logs, .. } => {
-                let topics: HashSet<B256> = logs
-                    .into_iter()
-                    .flat_map(|log| log.topics().to_vec())
-                    .collect();
-                if !expected_logs
-                    .iter()
-                    .all(|expected_topic| topics.contains(expected_topic))
-                {
-                    return Err(BuilderTransactionError::InvalidContract(
-                        to,
-                        InvalidContractDataError::InvalidLogs(
-                            expected_logs,
-                            topics.into_iter().collect(),
-                        ),
-                    ));
-                }
-                let return_output = T::abi_decode_returns(&output.into_data()).map_err(|_| {
-                    BuilderTransactionError::InvalidContract(
-                        to,
-                        InvalidContractDataError::OutputAbiDecodeError,
-                    )
-                })?;
-                Ok(SimulationSuccessResult::<T> {
-                    gas_used,
-                    output: return_output,
-                    state_changes: state,
-                })
-            }
-            ExecutionResult::Revert { output, .. } => {
-                let revert = ContractError::<E>::abi_decode(&output)
-                    .map(|reason| Revert::from(format!("{reason:?}")))
-                    .or_else(|_| Revert::abi_decode(&output))
-                    .unwrap_or_else(|_| {
-                        Revert::from(format!("unknown revert: {}", hex::encode(&output)))
-                    });
-                Err(BuilderTransactionError::TransactionReverted(to, revert))
-            }
-            ExecutionResult::Halt { reason, .. } => {
-                Err(BuilderTransactionError::TransactionHalted(to, reason))
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub(super) struct BuilderTxBase {
-    pub signer: Option<Signer>,
-}
-
-impl BuilderTxBase {
-    pub(super) fn new(signer: Option<Signer>) -> Self {
-        Self { signer }
-    }
-
-    pub(super) fn simulate_builder_tx(
-        &self,
-        ctx: &BuilderTxEnv<'_>,
-        db: impl DatabaseRef,
-    ) -> Result<Option<BuilderTransactionCtx>, BuilderTransactionError> {
-        match self.signer {
-            Some(signer) => {
-                let message: Vec<u8> = format!("Block Number: {}", ctx.block_number).into_bytes();
-                let gas_used = self.estimate_builder_tx_gas(&message);
-                let signed_tx = self.signed_builder_tx(ctx, db, signer, gas_used, message)?;
-                let da_size = op_alloy_flz::tx_estimated_size_fjord_bytes(
-                    signed_tx.encoded_2718().as_slice(),
-                );
-                Ok(Some(BuilderTransactionCtx {
-                    gas_used,
-                    da_size,
-                    signed_tx,
-                    is_top_of_block: false,
-                }))
-            }
-            None => Ok(None),
-        }
-    }
-
-    fn estimate_builder_tx_gas(&self, input: &[u8]) -> u64 {
-        // Count zero and non-zero bytes
-        let (zero_bytes, nonzero_bytes) = input.iter().fold((0, 0), |(zeros, nonzeros), &byte| {
-            if byte == 0 {
-                (zeros + 1, nonzeros)
-            } else {
-                (zeros, nonzeros + 1)
-            }
-        });
-
-        // Calculate gas cost (4 gas per zero byte, 16 gas per non-zero byte)
-        let zero_cost = zero_bytes * 4;
-        let nonzero_cost = nonzero_bytes * 16;
-
-        // Tx gas should be not less than floor gas https://eips.ethereum.org/EIPS/eip-7623
-        let tokens_in_calldata = zero_bytes + nonzero_bytes * 4;
-        let floor_gas = 21_000 + tokens_in_calldata * TOTAL_COST_FLOOR_PER_TOKEN;
-
-        std::cmp::max(zero_cost + nonzero_cost + 21_000, floor_gas)
-    }
-
-    fn signed_builder_tx(
-        &self,
-        ctx: &BuilderTxEnv<'_>,
-        db: impl DatabaseRef,
-        signer: Signer,
-        gas_used: u64,
-        message: Vec<u8>,
-    ) -> Result<Recovered<OpTransactionSigned>, BuilderTransactionError> {
-        let nonce = get_nonce(db, signer.address)?;
-
-        // Create the EIP-1559 transaction
-        let tx = OpTypedTransaction::Eip1559(TxEip1559 {
-            chain_id: ctx.chain_id(),
-            nonce,
-            gas_limit: gas_used,
-            max_fee_per_gas: ctx.base_fee.into(),
-            max_priority_fee_per_gas: 0,
-            to: TxKind::Call(Address::ZERO),
-            // Include the message as part of the transaction data
-            input: message.into(),
-            ..Default::default()
-        });
-        // Sign the transaction
-        let builder_tx = signer
-            .sign_tx(tx)
-            .map_err(BuilderTransactionError::SigningError)?;
-
-        Ok(builder_tx)
-    }
-}
-
-pub fn get_nonce(db: impl DatabaseRef, address: Address) -> Result<u64, BuilderTransactionError> {
-    db.basic_ref(address)
-        .map(|acc| acc.unwrap_or_default().nonce)
-        .map_err(|_| BuilderTransactionError::AccountLoadFailed(address))
-}
-
-pub fn get_balance(
-    db: impl DatabaseRef,
-    address: Address,
-) -> Result<U256, BuilderTransactionError> {
-    db.basic_ref(address)
-        .map(|acc| acc.unwrap_or_default().balance)
-        .map_err(|_| BuilderTransactionError::AccountLoadFailed(address))
-}
-
-/// Adjust batch gas/DA/uncompressed limits to reserve capacity for bottom-of-block builder txs.
-/// Returns the adjusted `max_uncompressed_block_size`.
-pub(crate) fn reserve_builder_tx_budget(
-    builder_txs: &[BuilderTransactionCtx],
-    target_gas: &mut u64,
-    target_da: &mut Option<u64>,
-    target_da_footprint: &mut Option<u64>,
-    da_footprint_scalar: Option<u16>,
-    ctx_max_uncompressed: Option<u64>,
-    cumulative_uncompressed: u64,
-) -> Option<u64> {
-    let bottom_txs = builder_txs.iter().filter(|tx| !tx.is_top_of_block);
-
-    let mut builder_tx_gas: u64 = 0;
-    let mut builder_tx_da_size: u64 = 0;
-    let mut builder_tx_uncompressed_size: u64 = 0;
-    for tx in bottom_txs {
-        builder_tx_gas += tx.gas_used;
-        builder_tx_da_size += tx.da_size;
-        builder_tx_uncompressed_size += tx.signed_tx.inner().encode_2718_len() as u64;
-    }
-
-    *target_gas = target_gas.saturating_sub(builder_tx_gas);
-
-    if let Some(da_limit) = target_da.as_mut() {
-        *da_limit = da_limit.saturating_sub(builder_tx_da_size);
-    }
-
-    if let (Some(footprint), Some(scalar)) = (target_da_footprint.as_mut(), da_footprint_scalar) {
-        *footprint = footprint.saturating_sub(builder_tx_da_size.saturating_mul(u64::from(scalar)));
-    }
-
-    let max_uncompressed =
-        ctx_max_uncompressed.map(|limit| limit.saturating_sub(builder_tx_uncompressed_size));
-    if let Some(limit) = ctx_max_uncompressed
-        && cumulative_uncompressed >= limit.saturating_sub(builder_tx_uncompressed_size)
-    {
-        error!(
-            target: "payload_builder",
-            current_uncompressed = cumulative_uncompressed,
-            reserved_builder_tx_uncompressed = builder_tx_uncompressed_size,
-            limit,
-            "Builder tx uncompressed size subtraction caused max_uncompressed_block_size to be 0. \
-             No transaction would be included."
-        );
-    }
-
-    max_uncompressed
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn mock_builder_tx(gas_used: u64, da_size: u64, top_of_block: bool) -> BuilderTransactionCtx {
-        let signer = Signer::random();
-        let signed = signer
-            .sign_tx(OpTypedTransaction::Legacy(Default::default()))
-            .unwrap();
-        BuilderTransactionCtx {
-            gas_used,
-            da_size,
-            signed_tx: signed,
-            is_top_of_block: top_of_block,
-        }
-    }
-
-    #[test]
-    fn test_reserve_budget_empty_txs() {
-        let mut gas = 1_000_000;
-        let mut da = Some(10_000u64);
-        let mut footprint = Some(50_000u64);
-
-        let result = reserve_builder_tx_budget(
-            &[],
-            &mut gas,
-            &mut da,
-            &mut footprint,
-            Some(1),
-            Some(100_000),
-            0,
-        );
-
-        assert_eq!(gas, 1_000_000, "gas unchanged");
-        assert_eq!(da, Some(10_000), "da unchanged");
-        assert_eq!(footprint, Some(50_000), "footprint unchanged");
-        assert_eq!(result, Some(100_000), "uncompressed unchanged");
-    }
-
-    #[test]
-    fn test_reserve_budget_only_bottom_txs_counted() {
-        let txs = vec![
-            mock_builder_tx(100_000, 500, true), // top-of-block — should be ignored
-            mock_builder_tx(50_000, 200, false), // bottom — counted
-        ];
-        let mut gas = 1_000_000;
-        let mut da = Some(10_000u64);
-        let mut footprint = None;
-
-        reserve_builder_tx_budget(&txs, &mut gas, &mut da, &mut footprint, None, None, 0);
-
-        assert_eq!(gas, 950_000, "only bottom tx gas subtracted");
-        assert_eq!(da, Some(9_800), "only bottom tx da subtracted");
-    }
-
-    #[test]
-    fn test_reserve_budget_saturating_sub() {
-        let txs = vec![mock_builder_tx(2_000_000, 50_000, false)];
-        let mut gas = 1_000_000;
-        let mut da = Some(10_000u64);
-        let mut footprint = Some(100u64);
-
-        reserve_builder_tx_budget(&txs, &mut gas, &mut da, &mut footprint, Some(1), None, 0);
-
-        assert_eq!(gas, 0, "gas saturates at 0");
-        assert_eq!(da, Some(0), "da saturates at 0");
-        assert_eq!(footprint, Some(0), "footprint saturates at 0");
-    }
-
-    #[test]
-    fn test_reserve_budget_da_none_untouched() {
-        let txs = vec![mock_builder_tx(100, 200, false)];
-        let mut gas = 1_000;
-        let mut da: Option<u64> = None;
-        let mut footprint: Option<u64> = None;
-
-        reserve_builder_tx_budget(&txs, &mut gas, &mut da, &mut footprint, None, None, 0);
-
-        assert_eq!(gas, 900);
-        assert_eq!(da, None, "None da stays None");
-        assert_eq!(footprint, None, "None footprint stays None");
-    }
-
-    #[test]
-    fn test_reserve_budget_uncompressed_limit_reduction() {
-        let txs = vec![mock_builder_tx(0, 0, false)];
-        // The tx has some encoded size even with default fields
-        let mut gas = 1_000_000;
-
-        let result = reserve_builder_tx_budget(
-            &txs,
-            &mut gas,
-            &mut None,
-            &mut None,
-            None,
-            Some(1_000_000),
-            0,
-        );
-
-        // Result should be Some(1_000_000 - encoded_size)
-        assert!(result.is_some());
-        assert!(
-            result.unwrap() < 1_000_000,
-            "uncompressed limit reduced by tx encoded size"
-        );
-    }
-}
