+++ crates/op-rbuilder/src/builder/builder_tx/impls/mod.rs
@@ -1,0 +1,5 @@
+mod claim;
+mod flashblock_number;
+
+pub(crate) use claim::ClaimBuilderTx;
+pub(crate) use flashblock_number::FlashblockNumberBuilderTx;
